Remove src/ndv/views/bases/_graphics
The few functions that still have use have been moved to
src/ndv/views/_util
