--- src/ndv/views/bases/_graphics/__init__.py
@@ -1,5 +1,0 @@
-"""Base classes for graphics elements."""
-
-# This should probably be deleted at some point.
-
-__all__ = []
